pyproject.toml
[project]
name = "streaming-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "faiss-cpu>=1.7.0",
    "python-dotenv>=1.1.1",
    "aiohttp>=3.8.0",
]
